separated user views to app/user
